kitchen/requirements.txt
grpcio-tools ~= 1.30
requests
simplejson
